refactor: HeroImage 컴포넌트 분리
