+++ apps/web/public/images/hero/2.jpg
